Actor: Pick smallest delta when moving away from block
This patch fixes a bug that would cause an actor to be moved to an
incorrect side of a block, when it has a constraint to never overlap a
block in a given group.

The fix is to calculate the smallest possible delta to move the actor by,
and pick that delta when moving the actor away. The result is that the
actor will now snap to the most logical edge of the block instead of being
moved to another one of the block’s sides.