Actor: Render first animation frame directly
This change makes actors render their initial animation frame directly
when the animation is assigned, rather than waiting for the next screen
update. The reason for this is to enable the actor’s size to be set as
soon as it’s added to a scene, to enable easier positioning.

This also enables an optimization path in that an animation action does
not have to be added in case only a single frame is being rendered.

This change also includes a new convenience API to create an animation
from an array of images, which enables more thorough testing of the
resizing feature.