Add support for sprite sheet-based animations
This change adds support for using a single sprite sheet texture for an
animation, that lets an actor animate over sub-rectangles of that texture,
which can contain a single or multiple rows.

This new feature is a compliment to the already existing option of using
separate image files for each texture. The advantage of using a sprite
sheet is that only a single image needs to be loaded into memory, and
drawing can be done by simply offsetting that single texture instead of
having to draw a completely new one.

This change very slightly breaks the existing `Animation` API in that an
array of textures need to be wrapped in a `.textures` enum case if assigned
directly to an existing animation.