Actor: Combine move events & remove duplicates
This fix merges the previously introduced `willMove` event into the `move`
event and makes it include a tuple with both the old and new positions as
its value. This enables logic that uses either the old or new position to
use the exact same event.

Further, a fix has been made to not send duplicate events when a position
is corrected by, for example, constraints. Only the final position is now
emitted as an event, if it differs from the original.