Scene: Add safe area insets API
Similar to how UIKit now includes a `safeAreaInsets` API to be able to
easily place UI elements away from the notch & home indicator on iPhone X.

Changes required:

- Add `safeAreaInsets` property on `Scene`, note that it’s not only available
on iOS, but on all platforms - preserving the cross platformness of Imagine.

- Add `safeAreaInsetsChanged` event on `SceneEventCollection` that gets triggered
when the safe area insets were changed.

- Make it possible to inject a mocked `GameView` in `Game`, to enable unit testing
of this new feature.

- Make a cross-platform type alias of `NS/UIEdgeInsets = EdgeInsets`.