TextureManager: Add support for assigning a scene-wide name prefix
This change adds a `namePrefix` property on `TextureManager`, that can
be used to assign a prefix to all textures loaded by that manager. This
is super useful when all graphics for a game are contained in a single
folder, and lets you skip referring to that folder every time you reference
a texture.