Actor: Remove use of defer to trigger property observers
This change removes the use of `defer` to trigger a property’s `didSet`
observation. While this currently works, it has been confirmed to be
because of a compiler bug that will will be fixed soon.

Instead of relying on defer, the methods dealing with changes in values
are now explicitly called during init instead.