SwiftLint: Enable most rules + fix codebase to match
A lot of SwiftLint rules have now been enabled. They were initially disabled
to enable SwiftLint integration, but now it’s time to enable them and fix all
warnings and errors that they result in.

Most fixes are removal of whitespace and shortening of lines, and at some places
local disabling of rules had to be inserted (when force casting has to be used).