EventCollection: Enable custom events to be created with a subject type
Currently, it’s only possible to make custom events with a subject type
if you also have a subject identifier. While this works for cases where
you want to make the observation conditional on the subject, for events
that only pass a subject when observed this is not sufficient. This change
fixes this by allowing any subject type to be used with a custom event.