Collision detection: Activate lazily, make more thorough & remove any events
This change introduces several optimizations in how collision detection
is performed for both actors and blocks, in order to avoid duplicate or
unnecessary operations.

- Activate an object’s collision detection lazily, whenever a collision
event is first accessed.
- Remove collision events that target any actor or any block. These make
it hard to perform optimizations, and are very unlikely to be used in any
real-life game development.
- Keep track of whenever a block and actor is in contact, to avoid handling
duplicate collisions.
- Don’t perform collision detection unless someone’s interested in it, and
in case we only need to do collision detection to evaluate constraints, there’s
now an optimization path for that.

With this change, adding 100 objects on the same point in a scene is now over
20 times faster!