Scene: Fix actors/labels incorrectly returned for a given point
This patch fixes a bug that could cause actors or labels to be incorrectly
returned when asking for objects at a certain point. The problem was that
all members of the grid tile in question would be returned. Now, we always
check that an object’s rect contains the given point.