Grid: Handle activation & deactivation
This is a refactor in preparation of an upcoming change that will enable
scene objects to be added to a `Camera`, as well as to a `Scene`. In order
to avoid code duplication, the goal is to move more object management
responsibilities into the `Grid` class, starting with activation & deactivation.