Refactor deactivation code to make it more predictable
This change refactors the way scene objects are deactivated and removed
from a scene. Currently, the concepts of deactivation and removal are a
bit entangled, so this change clears that up a bit.

- Deactivation: The object is no longer active in a scene, either because
  it was removed, or because the scene itself was deactivated.
- Removal: The object and its layer are removed from the scene.

To make this happen, `Grid` is now responsible for all deactivation, including
when the scene is reset or when an object is removed.

To increase predictability, the only way to remove scene objects are now by
calling `remove()` on the object itself, rather than being able to use an API
on `Scene`.