Camera: Add ‘clicked’ event
With this change `Camera` now has a `clicked` event that can be used to
observe clicks within the camera’s coordinate space. This change also
includes a refactor to make the click handling handled more by the
individual objects and by `Grid`, to make it easier to add more click
features in the future.