Introduce sprite sheet slicing
This change makes it possible to slice a sprite sheet up for animation
purposes. This enables the API user to use sprite sheet as atlases, with
many animations baked into the same sprite sheet.

Using a new slicing API on `SpriteSheet` combined with a new `Coordinate`
structure, you can now easily reference a “slice” of a sprite sheet when
creating an animation.

When implementing this, the frame calculation code for sprite sheets was
moved from `Animation` into `SpriteSheet` itself.