Actor: Rework contact management to support edge-to-edge contacts
This changes the way contacts are managed between actors and blocks, and
pushes the responsibility of managing object contacts back to `Grid`,
instead of being handled in each separate object.

The new implementation is more robust, since it invalidates contacts on
each rect change, rather than just when an object moves between different
grid tiles. It also now treats edge-to-edge contacts (that is, no overlap
but a distance of 1 pixel or less between two objects) as a contact. This
is really useful in platform games, when you want to check if the player
is currently standing on the ground, for example.